requirements.txt
keyboard
pillow
requests
pyautogui